+++ opensoundscape/datasets.py
@@ -1,3 +1,6 @@
+# TODO (enhancement): add MelSpectrogram, Scalogram, MFCC preprocessors
+# TODO: separate module for preprocessors? maybe folder with __init__ imports
+
 #!/usr/bin/env python3
 import pandas as pd
 import numpy as np
@@ -211,274 +214,3 @@ class SplitterDataset(torch.utils.data.Dataset):
     @classmethod
     def collate_fn(*batch):
         return chain.from_iterable([x["data"] for x in batch[1]])
-
-
-from opensoundscape import preprocess
-from opensoundscape.preprocess import ParameterRequiredError
-
-
-class BasePreprocessor(torch.utils.data.Dataset):
-    def __init__(self, df, return_labels=True):
-        # TODO: add .sample method?
-
-        assert Path(df.index[0]).exists(), (
-            "Index of dataframe passed to "
-            f"preprocessor must be a file path. Got {df.index[0]}."
-        )
-
-        self.df = df
-        self.return_labels = return_labels
-        self.labels = df.columns
-
-        # actions: a collection of instances of BaseAction child classes
-        self.actions = preprocess.ActionContainer()
-
-        # pipeline: an ordered list of operations to conduct on each file,
-        # each pulled from self.actions
-        self.pipeline = []
-
-    def __len__(self):
-        return self.df.shape[0]
-
-    def __getitem__(self, item_idx):
-
-        df_row = self.df.iloc[item_idx]
-        x = Path(df_row.name)  # the index contains a path to a file
-
-        for pipeline_element in self.pipeline:
-            if pipeline_element.bypass:
-                continue
-            try:
-                x = pipeline_element.go(x)
-            except ParameterRequiredError:  # need to pass labels
-                x, df_row = pipeline_element.go(x, df_row)
-
-        # Return sample & label pairs (training/validation)
-        if self.return_labels:
-            labels = torch.from_numpy(df_row.values)
-            return {"X": x, "y": labels}
-
-        # Return sample only (prediction)
-        return {"X": x}
-
-    def class_counts_cal(self):
-        """count number of each label"""
-        print("Warning: check if this is the correct behavior")
-        labels = self.df.columns
-        counts = np.sum(self.df.values, 0)
-        return labels, counts
-
-
-class AudioLoadingPreprocessor(BasePreprocessor):
-    """creates Audio objects from file paths"""
-
-    def __init__(self, df, return_labels=True):
-
-        super(AudioLoadingPreprocessor, self).__init__(df, return_labels=return_labels)
-
-        # add an AudioLoader to our (currently empty) action toolkit
-        self.actions.load_audio = preprocess.AudioLoader()
-
-        # add the action to our (currently empty) pipeline
-        self.pipeline.append(self.actions.load_audio)
-
-
-class AudioToImagePreprocessor(
-    BasePreprocessor
-):  # TODO: rename as AudioToSpectrogramPreprocessor
-    # TODO (enhancement): add MelSpectrogram, Scalogram, MFCC preprocessors
-    """loads audio paths, creates spectrogram, returns tensor
-
-    by default, resamples audio to sr=22050
-    can always change with .actions.load_audio.set(sample_rate=)
-
-    df must have audio paths as index and class names as columns
-    - if no labels, df will have no columns
-    """
-
-    def __init__(self, df, audio_length=None, out_shape=[224, 224], return_labels=True):
-
-        super(AudioToImagePreprocessor, self).__init__(df, return_labels=return_labels)
-
-        self.audio_length = audio_length
-        self.return_labels = return_labels
-
-        # add each action to our tool kit, then to pipeline
-        self.actions.load_audio = preprocess.AudioLoader(sample_rate=22050)
-        self.pipeline.append(self.actions.load_audio)
-
-        self.actions.trim_audio = preprocess.AudioTrimmer()
-        self.pipeline.append(self.actions.trim_audio)
-
-        self.actions.to_spec = preprocess.AudioToSpectrogram()
-        self.pipeline.append(self.actions.to_spec)
-
-        self.actions.bandpass = preprocess.SpectrogramBandpass(min_f=0, max_f=20000)
-        self.pipeline.append(self.actions.bandpass)
-        self.actions.bandpass.off()  # bandpass is off by default
-
-        self.actions.to_img = preprocess.SpecToImg(shape=out_shape)
-        self.pipeline.append(self.actions.to_img)
-
-        self.actions.to_tensor = preprocess.ImgToTensor()
-        self.pipeline.append(self.actions.to_tensor)
-
-        self.actions.normalize = preprocess.TensorNormalize()
-        self.pipeline.append(self.actions.normalize)
-
-
-class CnnPreprocessor(AudioToImagePreprocessor):
-    """Child of AudioToImagePreprocessor with full augmentation pipeline"""
-
-    def __init__(
-        self,
-        df,
-        audio_length=None,
-        return_labels=True,
-        debug=None,
-        overlay_df=None,
-        out_shape=[224, 224],
-    ):
-
-        super(CnnPreprocessor, self).__init__(
-            df,
-            audio_length=audio_length,
-            out_shape=out_shape,
-            return_labels=return_labels,
-        )
-
-        self.debug = debug
-
-        # extra Actions for augmentation steps
-        # TODO: should be able to create object without overlay_df
-        self.actions.overlay = preprocess.ImgOverlay(
-            overlay_df=overlay_df,
-            audio_length=self.audio_length,
-            overlay_prob=1,
-            max_overlay_num=1,
-            overlay_class=None,
-            # TODO: check - overlay pipeline might not update with changes?
-            loader_pipeline=self.pipeline[0:5],  # all actions before overlay
-            update_labels=False,
-        )
-
-        self.actions.color_jitter = preprocess.TorchColorJitter()
-        self.actions.random_affine = preprocess.TorchRandomAffine()
-        self.actions.time_mask = preprocess.TimeMask()
-        self.actions.frequency_mask = preprocess.FrequencyMask()
-        # self.actions.time_warp = preprocess.TimeWarp()
-        self.actions.add_noise = preprocess.TensorAddNoise(std=0.005)
-
-        self.augmentation_pipeline = [
-            self.actions.load_audio,
-            self.actions.trim_audio,
-            self.actions.to_spec,
-            self.actions.bandpass,
-            self.actions.to_img,
-            self.actions.overlay,
-            self.actions.color_jitter,
-            self.actions.to_tensor,
-            # self.actions.time_warp,
-            self.actions.time_mask,
-            self.actions.frequency_mask,
-            self.actions.add_noise,
-            self.actions.normalize,
-            self.actions.random_affine,
-        ]
-
-        self.no_augmentation_pipeline = [
-            self.actions.load_audio,
-            self.actions.trim_audio,
-            self.actions.to_spec,
-            self.actions.bandpass,
-            self.actions.to_img,
-            self.actions.to_tensor,
-            self.actions.normalize,
-        ]
-
-        if self.debug is not None:
-            self.actions.save_img = preprocess.SaveTensorToDisk(self.debug)
-            self.pipeline.append(self.actions.save_img)
-
-    def augmentation_on():
-        """use pipeline containing all actions including augmentations"""
-        self.pipeline = self.augmentation_pipeline
-
-    def augmentation_off():
-        """use pipeline that skips all augmentations"""
-        self.pipeline = self.no_augmentation_pipeline
-
-
-class ResnetMultilabelPreprocessor(BasePreprocessor):
-    """loads audio paths, performs various augmentations, returns tensor"""
-
-    def __init__(
-        self,
-        df,
-        audio_length=None,
-        return_labels=True,
-        augmentation=True,
-        debug=None,
-        overlay_df=None,
-    ):
-
-        super(ResnetMultilabelPreprocessor, self).__init__(
-            df, return_labels=return_labels
-        )
-
-        self.audio_length = audio_length
-        self.augmentation = augmentation
-        self.return_labels = return_labels
-        self.debug = debug
-
-        # add each action to our tool kit, then to pipeline
-        self.actions.load_audio = preprocess.AudioLoader()
-        self.pipeline.append(self.actions.load_audio)
-
-        self.actions.trim_audio = preprocess.AudioTrimmer()
-        self.pipeline.append(self.actions.trim_audio)
-
-        self.actions.to_spec = preprocess.AudioToSpectrogram()
-        self.pipeline.append(self.actions.to_spec)
-
-        self.actions.to_img = preprocess.SpecToImg()
-        self.pipeline.append(self.actions.to_img)
-
-        # should make one without overlay, then subclass and add overlay
-        if self.augmentation:
-            self.actions.overlay = preprocess.ImgOverlay(
-                overlay_df=overlay_df,
-                audio_length=self.audio_length,
-                overlay_prob=0.5,
-                max_overlay_num=2,
-                overlay_weight=[0.2, 0.5],
-                # this pipeline might not update with changes to preprocessor?
-                loader_pipeline=self.pipeline[0:4],
-                update_labels=False,
-            )
-            self.pipeline.append(self.actions.overlay)
-
-            # color jitter and affine can be applied to img or tensor
-            # here, we choose to apply them to the PIL.Image
-            self.actions.color_jitter = preprocess.TorchColorJitter()
-            self.pipeline.append(self.actions.color_jitter)
-
-            self.actions.random_affine = preprocess.TorchRandomAffine()
-            self.pipeline.append(self.actions.random_affine)
-
-        self.actions.to_tensor = preprocess.ImgToTensor()
-        self.pipeline.append(self.actions.to_tensor)
-
-        if self.augmentation:
-            self.actions.tensor_aug = preprocess.TensorAugment()
-            self.pipeline.append(self.actions.tensor_aug)
-
-            self.actions.add_noise = preprocess.TensorAddNoise(std=1.0)
-            self.pipeline.append(self.actions.add_noise)
-
-        self.actions.normalize = preprocess.TensorNormalize()
-        self.pipeline.append(self.actions.normalize)
-
-        if self.debug is not None:
-            self.actions.save_img = preprocess.SaveTensorToDisk(self.debug)
-            self.pipeline.append(self.actions.save_img)
