+++ opensoundscape/preprocess/tensor_augment.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python3
 """
-These functions were implemented for PyTorch in the following repository https://github.com/zcaceres/spec_augment
+Augmentations and transforms for torch.Tensors
+
+These functions were implemented for PyTorch in:
+https://github.com/zcaceres/spec_augment
 The original paper is available on https://arxiv.org/abs/1904.08779
 """
